Add Kore.Verified
