+++ kore/src/Kore/ASTVerifier/SentenceVerifier.hs
@@ -32,7 +32,7 @@ import qualified Kore.Builtin as Builtin
 import           Kore.Error
 import           Kore.IndexedModule.IndexedModule
 import           Kore.IndexedModule.Resolvers
-import qualified Kore.Verified.Sentence as Verified
+import qualified Kore.Verified as Verified
 
 {-|'verifyUniqueNames' verifies that names defined in a list of sentences are
 unique both within the list and outside, using the provided name set.
